pyproject.toml
[project]
name = "imc-image-converter"
version = "0.0.1"
description = "Fast, simple image format conversion for files."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["Pillow>=10.0.0"]

[project.scripts]
imc = "cli.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
